create_satellite_xml/requirements.txt
anyio==3.5.0
asgiref==3.5.0
beautifulsoup4==4.11.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Django==4.0.4
django-environ==0.8.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
flake8==4.0.1
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
lxml==4.8.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
rfc3986==1.5.0
sniffio==1.2.0
soupsieve==2.3.2
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
gunicorn